refactor: Piece 패키지 구조 변경
