+++ src/main/java/chess/model/piece/strategy/LimitedMovableStrategy.java
@@ -1,9 +1,9 @@
-package chess.model.strategy;
+package chess.model.piece.strategy;
 
 import chess.model.board.Square;
-import chess.model.strategy.move.Direction;
-import chess.model.strategy.move.Distance;
-import chess.model.strategy.move.MoveType;
+import chess.model.piece.strategy.move.Direction;
+import chess.model.piece.strategy.move.Distance;
+import chess.model.piece.strategy.move.MoveType;
 import java.util.List;
 
 public class LimitedMovableStrategy implements MovableStrategy {
