+++ src/main/java/chess/model/piece/strategy/move/Direction.java
@@ -1,4 +1,4 @@
-package chess.model.strategy.move;
+package chess.model.piece.strategy.move;
 
 import java.util.Arrays;
