+++ src/test/java/chess/model/piece/BishopTest.java
@@ -6,7 +6,7 @@ import chess.model.Color;
 import chess.model.File;
 import chess.model.Rank;
 import chess.model.board.Square;
-import chess.model.strategy.move.MoveType;
+import chess.model.piece.strategy.move.MoveType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
